[ADD] Migrated project_team module in v11.
