+++ project_team/__manifest__.py
@@ -8,15 +8,19 @@
     'summary': 'Project Team Management',
     'category': 'Project Management',
     'website': 'http://www.serpentcs.com',
-    'version': '10.0.1.0.0',
+    'version': '11.0.1.0.0',
     'license': 'AGPL-3',
-    'sequence': 1,
-    'depends': ['project', 'crm', 'web'],
+    'depends': [
+        'project',
+        'crm',
+        'web',
+    ],
     'data': [
-        'views/template.xml',
+#        'views/template.xml',
         'views/project_team_view.xml',
     ],
-    'images': ['static/description/ProjectTeam.png'],
+    'images': [
+        'static/description/ProjectTeam.png',
+    ],
     'installable': True,
-    'auto_install': False
 }
