+++ project_team/models/__init__.py
@@ -2,3 +2,4 @@
 # See LICENSE file for full copyright and licensing details.
 
 from . import project_team
+from . import crm_team
