+++ project_team/models/project_team.py
@@ -4,19 +4,6 @@
 from odoo import api, fields, models
 
 
-class CrmTeamInherit(models.Model):
-    _inherit = 'crm.team'
-
-    type_team = fields.Selection([('sale', 'Sale'), ('project', 'Project')],
-                                 string="Type", default="sale")
-
-    team_members = fields.Many2many('res.users', 'project_team_user_rel',
-                                    'team_id', 'uid', 'Project Members',
-                                    help="""Project's members are users who
-                                     can have an access to the tasks related
-                                     to this project.""")
-
-
 class ProjectProject(models.Model):
     _inherit = 'project.project'
 
@@ -25,7 +12,7 @@ class ProjectProject(models.Model):
                                members are users who can have an access to
                                the tasks related to this project."""
                                )
-    team_id = fields.Many2one('crm.team', string="Project Team",
+    team_id = fields.Many2one('crm.team', "Project Team",
                               domain=[('type_team', '=', 'project')])
 
     @api.onchange('team_id')
@@ -33,4 +20,4 @@ class ProjectProject(models.Model):
         self.members = False
         if self.team_id:
             self.members = \
-                [(6, 0, [rec.id for rec in self.team_id.team_members])]
+                [(6, 0, [member.id for member in self.team_id.team_members])]
