+++ project_team/views/project_team_view.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <odoo>
 
-    <record id="sales_team.sales_team_config_action"
-            model="ir.actions.act_window">
+    <record id="sales_team.sales_team_config_action" model="ir.actions.act_window">
         <field name="domain">[('type_team','=','sale')]</field>
         <field name="context">{'default_type_team':'sale'}</field>
     </record>
 
-    <!-- PROJECT TEAM FORM VIEW -->
     <record id="project_team_form_view" model="ir.ui.view">
         <field name="name">crm.team.form</field>
         <field name="model">crm.team</field>
@@ -16,18 +14,17 @@
                 <sheet>
                     <div class="oe_title">
                         <label for="name" class="oe_edit_only"
-                               string="Project team"/>
+                               string="Project Team"/>
                         <h1>
                             <field name="name"/>
                         </h1>
-                        <div name="options_active"></div>
+                        <div name="options_active"/>
                     </div>
                     <group>
                         <group name="left">
                             <field name="user_id"
                                    context="{'default_groups_ref': ['base.group_user', 'base.group_partner_manager', 'base.group_sale_salesman_all_leads']}"/>
                             <field name="type_team" invisible="1"/>
-                            <!-- <field name="code"/> -->
                         </group>
                     </group>
                     <notebook>
@@ -81,7 +78,6 @@
         </field>
     </record>
 
-    <!-- PROJECT TEAM TREE VIEW -->
     <record id="project_team_tree_view" model="ir.ui.view">
         <field name="name">crm.team.tree</field>
         <field name="model">crm.team</field>
@@ -93,7 +89,6 @@
         </field>
     </record>
 
-    <!-- CRM TEAM ACTION VIEW -->
     <record id="project_team_action_tree" model="ir.actions.act_window">
         <field name="name">Projects Team</field>
         <field name="res_model">crm.team</field>
@@ -111,8 +106,7 @@
         <field name="act_window_id" ref="project_team_action_tree"/>
     </record>
 
-    <record id="project_team_action_form_view"
-            model="ir.actions.act_window.view">
+    <record id="project_team_action_form_view" model="ir.actions.act_window.view">
         <field name="view_mode">form</field>
         <field name="view_id" ref="project_team_form_view"/>
         <field name="act_window_id" ref="project_team_action_tree"/>
@@ -122,59 +116,47 @@
     <menuitem id="project_team_menu" action="project_team_action_tree"
               parent="project.menu_project_config"/>
 
-    <record id="sales_team.crm_team_salesteams_act"
-            model="ir.actions.act_window">
+    <record id="sales_team.crm_team_salesteams_act" model="ir.actions.act_window">
         <field name="domain">[('type_team','=','sale')]</field>
     </record>
 
-    <!--INHERITED PROJECT FORM VIEW-->
     <record id="project_project_inherit_form_view" model="ir.ui.view">
         <field name="name">project.project.inherit.form.view</field>
         <field name="model">project.project</field>
         <field name="inherit_id" ref="project.edit_project"/>
         <field name="arch" type="xml">
             <field name="partner_id" position="after">
-                <newline/>
                 <field name="team_id"
                        context="{'form_view_ref': 'project_team.project_team_form_view', 'default_type_team':'project'}"/>
             </field>
-            <xpath expr="//group[@name='misc']/group" position="attributes">
-                <attribute name="colspan">4</attribute>
-                <attribute name="col">4</attribute>
-            </xpath>
-            <xpath expr="//page[1]" position="after">
+            <page name="settings" position="after">
                 <page string="Team">
-                    <group colspan="8">
-                        <field colspan="4" name="members"
-                               widget="many2many_kanban"
-                               nolabel="1"
-                               context="{'form_view_ref': 'project_team.project_team_form_view', 'default_type_team':'project'}">
-                            <kanban quick_create="false" create="true"
-                                    delete="true">
-                                <field name="id"/>
-                                <field name="image_small"/>
-                                <field name="name"/>
-                                <templates>
-                                    <t t-name="kanban-box">
-                                        <div style="position: relative">
-                                            <a t-if="!read_only_mode"
-                                               type="delete"
-                                               class="fa fa-times pull-right"/>
-                                            <div class="oe_module_vignette">
-                                                <img t-att-src="kanban_image('res.users', 'image_small', record.id.value)"
-                                                     class="oe_avatar oe_kanban_avatar_smallbox"/>
-                                                <div class="oe_module_desc">
-                                                    <field name="name"/>
-                                                </div>
+                    <field name="members" widget="many2many_kanban"
+                           context="{'form_view_ref': 'project_team.project_team_form_view', 'default_type_team':'project'}">
+                        <kanban quick_create="false" create="true" delete="true">
+                            <field name="id"/>
+                            <field name="image_small"/>
+                            <field name="name"/>
+                            <templates>
+                                <t t-name="kanban-box">
+                                    <div style="position: relative">
+                                        <a t-if="!read_only_mode" type="delete"
+                                           class="fa fa-times pull-right"/>
+                                        <div class="oe_module_vignette">
+                                            <img t-att-src="kanban_image('res.users', 'image_small', record.id.value)"
+                                                 class="oe_avatar oe_kanban_avatar_smallbox"/>
+                                            <div class="oe_module_desc">
+                                                <field name="name"/>
                                             </div>
                                         </div>
-                                    </t>
-                                </templates>
-                            </kanban>
-                        </field>
-                    </group>
+                                    </div>
+                                </t>
+                            </templates>
+                        </kanban>
+                    </field>
                 </page>
-            </xpath>
+            </page>
         </field>
     </record>
+
 </odoo>
